requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-pgvector==1.0.0
openai==1.3.2
torch==2.1.1
transformers==4.35.2